Kafka Consumer, Producer & controller method
KafkaConsumer, KafkaProducer classes have been added to the project.

In SeatsBookingSystemController publish method has been implemented.

Producer has been added to send String messages to the "kafkaTopic"
topic.

Consumer receives messages sent to the topic.

Probably this stuff will be changed in the future but it was implemented
for the kafka learning purpose.

Signed-off-by: Kamil Laganowski <laganowski.kamil@gmail.com>